+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Fix filter-only slots not receiving `options` attribute, causing select filters to show "No options available"
 * Ensure that field names have special characters stripped before being used in HTML attributes (#62)
+* Remove double-processing of select/multi-select options and ensure that falsy values are still processed (#63)
 
 ## v0.7.0 (2025-09-24)